refactor(iwfpapp): reorg test structure
