+++ iwfpapp/test/integration/cards_screen_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:iwfpapp/screens/cards/main.dart';
 import 'package:iwfpapp/services/data_store.dart';
-import 'mock_data_store.dart';
+import 'mock_services/mock_data_store.dart';
 
 void main() {
   testWidgets('test cards widget render no crash', (WidgetTester tester) async {
