+++ iwfpapp/test/integration/login_screen_test.dart
@@ -5,9 +5,9 @@ import 'package:iwfpapp/services/auth.dart';
 import 'package:iwfpapp/services/data_store.dart';
 import 'package:mockito/mockito.dart';
 import 'package:iwfpapp/services/mode.dart';
-import 'mock_mode.dart';
-import 'mock_data_store.dart';
-import 'mock_auth.dart';
+import 'mock_services/mock_mode.dart';
+import 'mock_services/mock_data_store.dart';
+import 'mock_services/mock_auth.dart';
 
 void main() {
   testWidgets('test contrib widget render no crash',
