+++ iwfpapp/test/integration/user_screen_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:iwfpapp/screens/user/main.dart';
 import 'package:iwfpapp/services/mode.dart';
-import './mock_mode.dart';
+import 'mock_services/mock_mode.dart';
 
 void main() {
   testWidgets('test user widget render no crash', (WidgetTester tester) async {
